% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidygtfs.R
\name{as_tidygtfs}
\alias{as_tidygtfs}
\title{Convert another gtfs like object to a tidygtfs object}
\usage{
as_tidygtfs(x, ...)
}
\arguments{
\item{x}{gtfs object}

\item{...}{ignored}
}
\value{
a tidygtfs object
}
\description{
Convert another gtfs like object to a tidygtfs object
}
