% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-gtfs-structure.R
\name{calendar_exception_fix}
\alias{calendar_exception_fix}
\title{checks for the missing calendar.txt exception (see https://developers.google.com/transit/gtfs/reference/calendar_dates-file). if the exception is TRUE, then calendar has its file spec set to 'extra'.}
\usage{
calendar_exception_fix(all_val_df)
}
\arguments{
\item{all_val_df}{dataframe of all files and fields. returned from validate_vars_provided()}
}
\value{
corrected dataframe.
}
\description{
checks for the missing calendar.txt exception (see https://developers.google.com/transit/gtfs/reference/calendar_dates-file). if the exception is TRUE, then calendar has its file spec set to 'extra'.
}
