% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-gtfs-helpers.R
\name{check_shape_ids}
\alias{check_shape_ids}
\title{Check whether shape ids were specified, or are not in the data, warning which ones were not found}
\usage{
check_shape_ids(gtfs_obj, shape_ids, shapes_routes_df)
}
\arguments{
\item{gtfs_obj}{a gtfs object}

\item{shape_ids}{a list of shape_ids to be plotted}

\item{shapes_routes_df}{shapes_routes_df - a dataframe in which routes and shapes are all cross referenced - can be used to filter shapes by service_id, for example}
}
\value{
a list of shape_ids that do exist in the data and can be plotted
}
\description{
Check whether shape ids were specified, or are not in the data, warning which ones were not found
}
