% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{convert_char_to_hms}
\alias{convert_char_to_hms}
\title{Convert time columns to hms::hms in feed}
\usage{
convert_char_to_hms(gtfs_list)
}
\arguments{
\item{gtfs_list}{gtfs object}
}
\value{
gtfs_obj with hms times columns for stop_times, frequencies and other
fields with type "Time"
}
\description{
Overwrites character columns in stop_times (arrival_time, departure_time) and
frequencies (start_time, end_time) with times converted with \code{\link[hms:hms]{hms::hms()}}.
}
\keyword{internal}
