% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-gtfs.R
\name{convert_gtfs_routes_to_sf}
\alias{convert_gtfs_routes_to_sf}
\title{return an sf dataframe with from a gtfs object}
\usage{
convert_gtfs_routes_to_sf(gtfs_obj)
}
\arguments{
\item{gtfs_obj}{gtfsr object}
}
\value{
an sf dataframe for gtfs routes with a multilinestring column
}
\description{
return an sf dataframe with from a gtfs object
}
