% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{convert_times_to_hms}
\alias{convert_times_to_hms}
\title{Use hms::hms columns in feed}
\usage{
convert_times_to_hms(gtfs_obj)
}
\arguments{
\item{gtfs_obj}{gtfs feed (tidygtfs object)}
}
\value{
gtfs_obj with added hms times columns for stop_times and frequencies
}
\description{
Overwirtes character columns in stop_times (arrival_time, departure_time) and
frequencies (start_time, end_time) with times converted with \code{\link[hms:hms]{hms::hms()}}.
}
