% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.R
\name{count_service_trips}
\alias{count_service_trips}
\title{Summarise the number of trips per service}
\usage{
count_service_trips(trips)
}
\arguments{
\item{trips}{trips data frame}
}
\value{
count of service by id
}
\description{
Summarise the number of trips per service
}
\keyword{internal}
