% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{download_from_url}
\alias{download_from_url}
\title{Download a zipped GTFS feed file from a url}
\usage{
download_from_url(url, path = tempfile(fileext = ".zip"), quiet = FALSE)
}
\arguments{
\item{url}{Character URL of GTFS feed.}

\item{path}{Character. Folder into which to put zipped file. If NULL, then save a tempfile}

\item{quiet}{Boolean. Whether to see file download progress. FALSE by default.}
}
\value{
File path
}
\description{
Download a zipped GTFS feed file from a url
}
\keyword{internal}
