% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_gtfs.R
\name{duplicated_unique_ids}
\alias{duplicated_unique_ids}
\title{Check if primary keys are unique within tables}
\usage{
duplicated_unique_ids(gtfs_list)
}
\arguments{
\item{gtfs_list}{list of tables}
}
\description{
Check if primary keys are unique within tables
}
