% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{empty_strings_to_na}
\alias{empty_strings_to_na}
\title{Convert empty strings ("") to NA values in gtfs tables}
\usage{
empty_strings_to_na(gtfs_obj)
}
\arguments{
\item{gtfs_obj}{gtfs feed (tidygtfs object)}
}
\value{
a gtfs_obj where all empty strings in tables have been replaced with NA
}
\description{
Convert empty strings ("") to NA values in gtfs tables
}
