% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins.R
\name{filter_feed_by_area}
\alias{filter_feed_by_area}
\title{Filter a gtfs feed so that it only contains trips that pass the given area}
\usage{
filter_feed_by_area(gtfs_obj, area)
}
\arguments{
\item{gtfs_obj}{tidygtfs object}

\item{area}{all trips passing through this area are kept. Either a bounding box
(numeric vector with xmin, ymin, xmax, ymax) or a sf object.}
}
\description{
Only keeps trips that pass stops within the given through_area.
Only stops, routes, services, shapes, frequencies and transfers belonging to
one of those trips are kept.
}
