% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins.R
\name{filter_feed_by_stops}
\alias{filter_feed_by_stops}
\title{Filter a gtfs feed so that it only contains trips that pass the given stops}
\usage{
filter_feed_by_stops(gtfs_obj, stop_ids = NULL, stop_names = NULL)
}
\arguments{
\item{gtfs_obj}{tidygtfs object}

\item{stop_ids}{vector with stop_ids}

\item{stop_names}{vector with stop_names}
}
\description{
Only keeps trips that pass stops with the given stop_ids or stop_names.
Only stops, routes, services, shapes, frequencies and transfers belonging to
one of those trips are kept. You can either provide stop_ids or stop_names
(which are then converted to stop_ids)
}
\details{
Note that the returned gtfs_obj contains more than just the stops given (i.e. all stops
that belong to a trip).
}
