% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_feed_by_trips}
\alias{filter_feed_by_trips}
\title{Filter a gtfs feed so that it only contains a given set of trips}
\usage{
filter_feed_by_trips(gtfs_obj, trip_ids)
}
\arguments{
\item{gtfs_obj}{gtfs feed (tidygtfs object)}

\item{trip_ids}{vector with trip_ids}
}
\value{
tidygtfs object with filtered tables
}
\description{
Only stop_times, stops, routes, services (in calendar and calendar_dates), shapes,
frequencies and transfers belonging to one of those trips are kept.
}
\seealso{
\code{\link{filter_feed_by_stops}}, \code{\link{filter_feed_by_area}}, \code{\link{filter_feed_by_date}}
}
