% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get-gtfs-feeds.R
\name{filter_feedlist}
\alias{filter_feedlist}
\title{Filter a feedlist to include only valid urls (ending in .zip)}
\usage{
filter_feedlist(feedlist_df)
}
\arguments{
\item{feedlist_df}{A dataframe of feed metadata such as output from get_feedlist}
}
\value{
A dataframe of feed metadata for all feeds in input that are downloadable
}
\description{
Filter a feedlist to include only valid urls (ending in .zip)
}

