% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel_times.R
\name{filter_stop_times}
\alias{filter_stop_times}
\title{Filter a \code{stop_times} table for a given date and timespan.}
\usage{
filter_stop_times(gtfs_obj, extract_date, min_departure_time, max_arrival_time)
}
\arguments{
\item{gtfs_obj}{gtfs feed (tidygtfs object)}

\item{extract_date}{date to extract trips from this day (Date or "YYYY-MM-DD" string)}

\item{min_departure_time}{(optional) The earliest departure time. Can be given as "HH:MM:SS",
hms object or numeric value in seconds.}

\item{max_arrival_time}{(optional) The latest arrival time. Can be given as "HH:MM:SS",
hms object or numeric value in seconds.}
}
\value{
Filtered \code{stop_times} data.table for \code{\link[=travel_times]{travel_times()}} and \code{\link[=raptor]{raptor()}}.
}
\description{
Filter a \code{stop_times} table for a given date and timespan.
}
\examples{
feed_path <- system.file("extdata", "routing.zip", package = "tidytransit")
g <- read_gtfs(feed_path)

# filter the sample feed
stop_times <- filter_stop_times(g, "2018-10-01", "06:00:00", "08:00:00")
}
