% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{filter_stop_times_by_hour}
\alias{filter_stop_times_by_hour}
\title{Filter stop times by hour of the day}
\usage{
filter_stop_times_by_hour(stop_times, start_hour, end_hour)
}
\arguments{
\item{stop_times_df}{a gtfsr$stop_times_df dataframe with lubridate arrival_time and departure_time}
}
\value{
dataframe with only stop times within the hours specified, with time columns as lubridate periods
}
\description{
Filter stop times by hour of the day
}
\keyword{internal}
