% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins.R
\name{filter_trips}
\alias{filter_trips}
\title{Filter all trips that pass the given stops}
\usage{
filter_trips(gtfs_obj, stop_ids = NULL, stop_names = NULL)
}
\arguments{
\item{gtfs_obj}{tidygtfs object}

\item{stop_ids}{vector with stop_ids}

\item{stop_names}{vector with stop_names}
}
\description{
Only keeps trips that pass stops with the given stop_ids or stop_names.
Only stops, routes, services, shapes, frequencies and transfers belonging to
one of those trips are kept.
You can either provide stop_ids or stop_names (which are then converted to stop_ids)
}
