% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-gtfs-networks.R
\name{get_agency_stops}
\alias{get_agency_stops}
\title{get stops for all routes of an agency}
\usage{
get_agency_stops(gtfs_obj, agency_id)
}
\arguments{
\item{gtfs_obj}{A GTFS list object with components agency_df, etc.}

\item{agency_id}{Character. Provide the ID of the agency whose routes are being mapped.}
}
\value{
dataframe of route and stop ids for given agency id
}
\description{
get stops for all routes of an agency
}

