% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gtfs.R
\name{get_feed}
\alias{get_feed}
\title{Download a zipped GTFS feed file from a url}
\usage{
get_feed(url, dir = "temp")
}
\arguments{
\item{url}{Character URL of GTFS feed.}

\item{dir}{Folder into which to put zipped file}
}
\value{
File path
}
\description{
Download a zipped GTFS feed file from a url
}

