% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-gtfs-feeds.R
\name{get_feedlist}
\alias{get_feedlist}
\title{Get list of all available feeds from transitfeeds API}
\usage{
get_feedlist()
}
\value{
a data frame with the result of httr::GET
}
\description{
Get list of all available feeds from transitfeeds API
}
