% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gtfs.R
\name{get_gtfs}
\alias{get_gtfs}
\title{Get a Dataframes of GTFS data.}
\usage{
get_gtfs(urls, timeout = 10, ...)
}
\arguments{
\item{path}{Character. url link to zip file}
}
\value{
Dataframes of GTFS data.
}
\description{
Get a Dataframes of GTFS data.
}

