% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-gtfs-files.R
\name{get_gtfs_meta}
\alias{get_gtfs_meta}
\title{Function that creates gtfs meta data}
\usage{
get_gtfs_meta()
}
\value{
Environment with gtfs data.
}
\description{
Function that creates gtfs meta data
}

