% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gtfs.R
\name{get_locations}
\alias{get_locations}
\title{Get all locations available from the transitfeeds API (getLocations)}
\usage{
get_locations()
}
\value{
A dataframe of locations with id, descriptions, and lat/lng
}
\description{
Get all locations available from the transitfeeds API (getLocations)
}

