% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-gtfs-helpers.R
\name{get_routes_sldf}
\alias{get_routes_sldf}
\title{Get shapes spatial data for given route ids}
\usage{
get_routes_sldf(gtfs_obj, route_ids, service_ids = NULL, shape_ids = NULL,
  route_opacity = NULL, route_colors = NULL)
}
\arguments{
\item{gtfs_obj}{A GTFS list object with components agency_df, etc.}

\item{route_ids}{Vector (Character). IDs for routes of interest.}

\item{service_ids}{Vector (Character). Service IDs. NULL by Default.}

\item{shape_ids}{Vector (Character). Shape IDs. NULL by Default.}

\item{route_opacity}{Numeric. Value must be between 0 and 1. Default is NULL.}

\item{route_colors}{Character. Names of colors (e.g. "blue") or hex values (e.g. '#000000').}
}
\value{
Environment containing spatial data, labels, colorings used for plotting
}
\description{
Get shapes spatial data for given route ids
}
