% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{get_shapes_geometry}
\alias{get_shapes_geometry}
\title{Make shapes into Simple Features Linestrings}
\usage{
get_shapes_geometry(shapes)
}
\arguments{
\item{shapes}{a gtfs$shapes dataframe}
}
\value{
an sf dataframe for gtfs shapes
}
\description{
Make shapes into Simple Features Linestrings
}
