% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{get_stops_geometry}
\alias{get_stops_geometry}
\title{Convert stops into Simple Features Points}
\usage{
get_stops_geometry(stops)
}
\arguments{
\item{stops}{a gtfs$stops dataframe}
}
\value{
an sf dataframe for gtfs routes with a point column
}
\description{
Convert stops into Simple Features Points
}
\examples{
data(gtfs_duke)
some_stops <- gtfs_duke$stops[sample(nrow(gtfs_duke$stops), 40),]
some_stops_sf <- get_stops_geometry(some_stops)
plot(some_stops_sf)
}
