% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtfs_duke}
\alias{gtfs_duke}
\title{Example GTFS data}
\format{
An object of class \code{tidygtfs} (inherits from \code{gtfs}) of length 25.
}
\usage{
gtfs_duke
}
\description{
Data obtained from
\url{https://data.trilliumtransit.com/gtfs/duke-nc-us/duke-nc-us.zip}.
}
\seealso{
read_gtfs
}
\keyword{datasets}
