% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtfs_obj}
\alias{gtfs_obj}
\title{Example GTFS data}
\format{An object of class \code{gtfs} of length 22.}
\usage{
gtfs_obj
}
\description{
Data obtained from
\url{http://data.trilliumtransit.com/gtfs/duke-nc-us/duke-nc-us.zip}.
}
\seealso{
read_gtfs
}
\keyword{datasets}
