% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{gtfs_route_to_sf_multilinestring}
\alias{gtfs_route_to_sf_multilinestring}
\title{return an sf multilinestring with lat and long from gtfs for a route}
\usage{
gtfs_route_to_sf_multilinestring(df)
}
\arguments{
\item{df}{dataframe with the shapes for a given route}
}
\value{
multilinestring (sfr) object
}
\description{
return an sf multilinestring with lat and long from gtfs for a route
}
