% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{gtfs_routes_as_sf}
\alias{gtfs_routes_as_sf}
\title{return an sf dataframe with from a gtfs object}
\usage{
gtfs_routes_as_sf(gtfs_obj)
}
\arguments{
\item{gtfs_obj}{gtfsr object}
}
\value{
an sf dataframe for gtfs routes with a multilinestring column
}
\description{
return an sf dataframe with from a gtfs object
}
