% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{gtfsr-package}
\alias{gtfsr}
\alias{gtfsr-package}
\title{A package for plotting GTFS data.}
\description{
This package allows one to quickly map GTFS data. It also provides a quick and simple way to validate the structure of GTFS data.
}
\examples{

 library(dplyr)
 url <- "http://data.trilliumtransit.com/gtfs/duke-nc-us/duke-nc-us.zip"
 gtfs_obj <- url \%>\% import_gtfs(quiet=TRUE)
 route <- "1693"
 map_gtfs(gtfs_obj, route)
}
\author{
Danton Noriega-Goodwin \email{danton.noriega@gmail.com},
Elaine McVey \email{elaine@transloc.com}
}

