% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get-gtfs-feeds.R
\name{has_api_key}
\alias{has_api_key}
\title{Make sure API key string is not empty}
\usage{
has_api_key()
}
\value{
logical TRUE if key is not empty
}
\description{
Make sure API key string is not empty
}

