% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{hhmmss_to_hms}
\alias{hhmmss_to_hms}
\title{convert a vector of time strings
empty strings are converted to NA}
\usage{
hhmmss_to_hms(time_strings)
}
\arguments{
\item{time_strings}{char vector ("HH:MM:SS")}
}
\description{
convert a vector of time strings
empty strings are converted to NA
}
\keyword{internal}
