% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{hhmmss_to_sec_split}
\alias{hhmmss_to_sec_split}
\title{Fallback function to convert strings like 5:02:11
10x slower than \code{\link[=hhmmss_to_seconds]{hhmmss_to_seconds()}}, empty strings are converted to NA}
\usage{
hhmmss_to_sec_split(hhmmss_str)
}
\arguments{
\item{hhmmss_str}{string}
}
\description{
Fallback function to convert strings like 5:02:11
10x slower than \code{\link[=hhmmss_to_seconds]{hhmmss_to_seconds()}}, empty strings are converted to NA
}
\keyword{internal}
