% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{hhmmss_to_seconds}
\alias{hhmmss_to_seconds}
\title{Convert "HH:MM:SS" time strings to seconds (numeric)
empty strings are converted to NA}
\usage{
hhmmss_to_seconds(hhmmss_str)
}
\arguments{
\item{hhmmss_str}{char vector ("HH:MM:SS")}
}
\description{
Convert "HH:MM:SS" time strings to seconds (numeric)
empty strings are converted to NA
}
\keyword{internal}
