% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-gtfs.R
\name{import_gtfs}
\alias{import_gtfs}
\title{Get a Dataframes of GTFS data.}
\usage{
import_gtfs(paths, local = FALSE, quiet = FALSE)
}
\arguments{
\item{paths}{Character. url links to zip files OR paths to local zip files. if to local path, then option `local` must be set to TRUE.}

\item{local}{Boolean. If the paths are searching locally or not. Default is FALSE (that is, urls).}

\item{quiet}{Boolean. Whether to see file download progress and files extract. FALSE by default.}
}
\value{
Dataframes of GTFS data.
}
\description{
Get a Dataframes of GTFS data.
}
