% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{join_gtfsr_dfs}
\alias{join_gtfsr_dfs}
\title{Merge gtfsr data frames across gtfsr objects}
\usage{
join_gtfsr_dfs(gtfs_obj_list, dfname)
}
\arguments{
\item{gtfs_obj_list}{a list of standard gtfsr objects}

\item{dfname}{dataframe to return}
}
\value{
one gtfsr dataframe object
}
\description{
merges gtfsr objects
}
