% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-gtfs.R
\name{join_shape_route_service_ids}
\alias{join_shape_route_service_ids}
\title{Join the shapes, trips and routes tables together - also checks on some potential errors in the data and warns accordingly}
\usage{
join_shape_route_service_ids(gtfs_obj, route_ids = NULL, service_ids = NULL)
}
\arguments{
\item{gtfs_obj}{a gtfs object}

\item{route_ids}{the routes for which to join the tables together - required, but not sure why this can't just be any/all routes in routes_df}

\item{service_ids}{- an optional filter for a certain service-default NULL}
}
\value{
shapes_routes_df - a dataframe in which routes,
# and shapes are all cross referenced - can be used to filter shapes by service_id, for example
}
\description{
Join the shapes, trips and routes tables together - also checks on some potential errors in the data and warns accordingly
}
