% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{list_files}
\alias{list_files}
\title{Read files with a "txt" suffix in a folder into objects in memory and delete files}
\usage{
list_files(ex_dir, delete_files = FALSE, quiet = FALSE)
}
\arguments{
\item{ex_dir}{Character. Path to folder into which files were extracted.}

\item{delete_files}{Logical, whether to delete the files after extraction.  Does not delete by default.}

\item{quiet}{Boolean. Whether to output messages and files found in folder.}
}
\description{
Read files with a "txt" suffix in a folder into objects in memory and delete files
}
\keyword{internal}
