% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-gtfs.R
\name{map_gtfs_all_stops_in_route}
\alias{map_gtfs_all_stops_in_route}
\title{map all stops for a route}
\usage{
map_gtfs_all_stops_in_route(gtfs_obj, route_id)
}
\arguments{
\item{gtfs_obj}{A GTFS list object with components agency_df, etc.}

\item{route_id}{Character. A single ID for a route of interest.}
}
\value{
Leaflet map object with all stop lat/long values plotted for a route.
}
\description{
map all stops for a route
}

