% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps-gtfs-routes.R
\name{map_gtfs_route_shape}
\alias{map_gtfs_route_shape}
\title{map the shape, with stops, for one or more routes}
\usage{
map_gtfs_route_shape(gtfs_obj, route_ids, service_ids = NULL,
  include_stops = TRUE)
}
\arguments{
\item{gtfs_obj}{A GTFS list object with components agency_df, etc.}

\item{route_ids}{Vector (Character). IDs for routes of interest.}

\item{service_ids}{Vector (Character). Service IDs. NULL by Default.}

\item{include_stops}{Boolean. Whether to layer on stops to the route shape. Default is TRUE.}
}
\value{
Leaflet map object with all stop lat/long values plotted for a route.
}
\description{
map the shape, with stops, for one or more routes
}

