% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps-gtfs-routes.R
\name{map_gtfs_routes}
\alias{map_gtfs_routes}
\title{map the shape, with stops, for one or more routes.}
\usage{
map_gtfs_routes(gtfs_obj, route_ids, service_ids = NULL,
  include_stops = TRUE, only_stops = FALSE, stop_opacity = 0.5,
  route_colors = NULL)
}
\arguments{
\item{gtfs_obj}{A GTFS list object with components agency_df, etc.}

\item{route_ids}{Vector (Character). IDs for routes of interest.}

\item{service_ids}{Vector (Character). Service IDs. NULL by Default.}

\item{include_stops}{Boolean. Whether to layer on stops to the route shape. Default is TRUE.}

\item{only_stops}{Boolean. Whether to map only stops, no routes. Overrides `include_stops`. Default is FALSE.}

\item{stop_opacity}{Numeric. Value must be between 0 and 1. Defaults is 0.5.}

\item{route_colors}{Character. Names of colors (e.g. "blue") or hex values (e.g. '#000000'). Default is NULL.}
}
\value{
Leaflet map object with all stop lat/long values plotted for a route.
}
\description{
map the shape, with stops, for one or more routes.
}

