% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-gtfs-stop.R
\name{map_gtfs_stop}
\alias{map_gtfs_stop}
\title{map a single stop}
\usage{
map_gtfs_stop(gtfs_obj, stop_id, stop_color = NULL)
}
\arguments{
\item{gtfs_obj}{A GTFS list object with components agency_df, etc.}

\item{stop_id}{Character. A single ID for a stop of interest.}

\item{stop_color}{Character. An R color or hex value. Expects single value or NULL. Default is NULL. If length(stop_color) > 1, then it reverts to `stop_color <- stop_color[1]`}
}
\value{
Leaflet map object with point plotted at stop lat/long value
}
\description{
map a single stop
}

