% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.R
\name{most_frequent_service}
\alias{most_frequent_service}
\title{Get the most frequent service for a set of trips.}
\usage{
most_frequent_service(trips)
}
\arguments{
\item{trips}{gtfs dataframe with a service_trips count column grouped by the service_id, route, etc of interest}
}
\value{
trips gtfs data frame within groups that are in the most frequent service window
}
\description{
Get the most frequent service for a set of trips.
}
\keyword{internal}
