% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{na_to_empty_strings}
\alias{na_to_empty_strings}
\title{Convert NA values to empty strings ("")}
\usage{
na_to_empty_strings(gtfs_obj)
}
\arguments{
\item{gtfs_obj}{tidygtfs object}
}
\description{
Convert NA values to empty strings ("")
}
