% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problems.R
\name{new_problems}
\alias{new_problems}
\title{Define new_problems environment, which contains functions used in validating GTFS data.}
\usage{
new_problems(gtfs_obj = NULL)
}
\description{
Define new_problems environment, which contains functions used in validating GTFS data.
}

