% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-gtfs-files.R
\name{parse_gtfs}
\alias{parse_gtfs}
\title{Function to better read in GTFS txt files}
\usage{
parse_gtfs(prefix, file_path)
}
\arguments{
\item{prefix}{Character. gtfs file prefix (e.g. 'agency', 'stop_times', etc.)}

\item{file_path}{Character. file path}
}
\description{
Function to better read in GTFS txt files
}

