% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{parse_hms_strings}
\alias{parse_hms_strings}
\title{convert a vector of time strings
empty strings are converted to NA}
\usage{
parse_hms_strings(time_strings)
}
\arguments{
\item{time_strings}{char vector ("HH:MM:SS")}
}
\description{
convert a vector of time strings
empty strings are converted to NA
}
