% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gtfs.R
\name{parse_stop_times}
\alias{parse_stop_times}
\title{Function to better read in stop_times.txt, which often fails}
\usage{
parse_stop_times(file_path, all_char = FALSE)
}
\arguments{
\item{file_path}{Character. file path}

\item{all_char}{Boolean. default FALSE. temporary variable. option to import all stop time data as character.}
}
\description{
Function to better read in stop_times.txt, which often fails
}

