% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{planner_buffer}
\alias{planner_buffer}
\title{Buffer using common urban planner distances}
\usage{
planner_buffer(df_sf1, dist = "h", crs = 26910)
}
\arguments{
\item{df_sf1}{a simple features data frame}

\item{dist}{default "h" - for half mile buffers. can also pass "q".}

\item{crs}{default epsg 26910. can be any other epsg}
}
\value{
a simple features data frame with planner buffers
}
\description{
merges gtfs objects
}
\keyword{internal}
