% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tidygtfs}
\alias{plot.tidygtfs}
\title{Plot GTFS stops and trips}
\usage{
\method{plot}{tidygtfs}(x, ...)
}
\arguments{
\item{x}{a tidygtfs object as read by \code{\link[=read_gtfs]{read_gtfs()}}}

\item{...}{ignored for tidygtfs}
}
\value{
plot
}
\description{
Plot GTFS stops and trips
}
\examples{
\donttest{
local_gtfs_path <- system.file("extdata",
                              "nyc_subway.zip",
                              package = "tidytransit")
nyc <- read_gtfs(local_gtfs_path)
plot(nyc)
}

}
