% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problems.R
\docType{data}
\name{problems}
\alias{problems}
\title{Define problems environment}
\format{An object of class \code{environment} of length 0.}
\usage{
problems
}
\description{
Define problems environment
}
\keyword{datasets}

