% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gtfs.R
\name{read_sub_gtfs}
\alias{read_sub_gtfs}
\title{Function to read all files into dataframes}
\usage{
read_sub_gtfs(file_path, prefix = "data/gtfs/", assign_envir = .GlobalEnv)
}
\arguments{
\item{file_path}{Character file path}
}
\description{
Function to read all files into dataframes
}

