% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmfolder}
\alias{rmfolder}
\title{Used to remove directory and its content}
\usage{
rmfolder(folder)
}
\arguments{
\item{folder}{Character. Path to folder.}
}
\description{
Used to remove directory and its content
}

