% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{route_type_names}
\alias{route_type_names}
\title{Dataframe of route type id's and the names of the types (e.g. "Bus")}
\format{
A data frame with 136 rows and 2 variables:
\describe{
\item{route_type}{the id of route type}
\item{route_type_name}{name of the gtfs route type}
}
}
\source{
\url{https://gist.github.com/derhuerst/b0243339e22c310bee2386388151e11e}
}
\usage{
route_type_names
}
\description{
Extended GTFS Route Types: https://developers.google.com/transit/gtfs/reference/extended-route-types
}
\keyword{datasets}
