% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{route_type_names_df}
\alias{route_type_names_df}
\title{Dataframe of route type id's and the names of the types (e.g. "Cable Car")}
\format{A data frame with 122 rows and 2 variables:
\describe{
\item{id}{the id of route type}
\item{name}{name of the gtfs route type}
}}
\source{
\url{https://gist.github.com/derhuerst/b0243339e22c310bee2386388151e11e}
}
\usage{
route_type_names_df
}
\description{
Dataframe of route type id's and the names of the types (e.g. "Cable Car")
}
\keyword{datasets}
