% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{routes_df_as_sf}
\alias{routes_df_as_sf}
\title{This function is deprecated. Please use get_route_geometry}
\usage{
routes_df_as_sf(gtfs_obj, route_ids = NULL, service_ids = NULL)
}
\description{
This function is deprecated. Please use get_route_geometry
}
