% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.R
\name{service_by_dow}
\alias{service_by_dow}
\title{Filter a gtfs calendar dataframe to service ids for specific days of the week}
\usage{
service_by_dow(calendar, dow = c(1, 1, 1, 1, 1, 0, 0))
}
\arguments{
\item{dow}{default to "weekday" (1,1,1,1,1,0,0)}

\item{gtfs_object}{object made by join_all_gtfs_tables}
}
\value{
service ids that match the schedule specified
}
\description{
Filter a gtfs calendar dataframe to service ids for specific days of the week
}
\keyword{internal}
