% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitfeeds.R
\name{set_api_key}
\alias{set_api_key}
\title{Set TransitFeeds API key for recall}
\usage{
set_api_key()
}
\value{
No value returned, function is used for setting environment variables
}
\description{
Set TransitFeeds API key for recall
}
