% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{set_hms_times}
\alias{set_hms_times}
\title{Use hms::hms columns in feed}
\usage{
set_hms_times(gtfs_obj)
}
\arguments{
\item{gtfs_obj}{a gtfs object in which hms times should be set,
the modified gtfs_obj is returned}
}
\value{
gtfs_obj with added hms times columns for stop_times and frequencies
}
\description{
Overwirtes character columns in stop_times (arrival_time, departure_time) and
frequencies (start_time, end_time) with times converted with \code{\link[hms:hms]{hms::hms()}}.
}
