% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.R
\name{set_servicepatterns}
\alias{set_servicepatterns}
\title{Calculate servicepatterns for the gtfs_obj}
\usage{
set_servicepatterns(gtfs_obj)
}
\arguments{
\item{gtfs_obj}{feed}
}
\value{
modified gtfs_obj with added servicepatterns list and a table linking trips and patterns (trip_servicepatterns)
}
\description{
Calculate servicepatterns for the gtfs_obj
}
\keyword{internal}
