% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trips.R
\name{set_trippattern}
\alias{set_trippattern}
\title{Add trip pattern data frame to the gtfs object. Trips (in stop_times) with the same
set of stops, arrival time and departure time have the same trip pattern.}
\usage{
set_trippattern(gtfs_obj, id_prefix = "t_", hash_algo = "md5", hash_length = 8)
}
\arguments{
\item{gtfs_obj}{gtfs feed}

\item{id_prefix}{all ids start with this string}

\item{hash_algo}{hashing algorithm used by digest}

\item{hash_length}{length the hash should be cut to with substr(). Use -1 if the full hash should be used}
}
\value{
gtfs_obj
}
\description{
Add trip pattern data frame to the gtfs object. Trips (in stop_times) with the same
set of stops, arrival time and departure time have the same trip pattern.
}
