% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{sf_as_tbl}
\alias{sf_as_tbl}
\title{Convert stops and shapes from sf objects to tibbles}
\usage{
sf_as_tbl(gtfs_obj)
}
\arguments{
\item{gtfs_obj}{gtfs feed (tidygtfs object)}
}
\value{
tidygtfs object with stops and shapes converted to tibbles
}
\description{
Coordinates are transformed to lon/lat columns (\code{stop_lon}/\code{stop_lat} or
\code{shape_pt_lon}/\code{shape_pt_lat})
}
\seealso{
\code{\link{gtfs_as_sf}}
}
