% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{sf_lines_to_df}
\alias{sf_lines_to_df}
\title{Adds the coordinates of an sf LINESTRING object as columns and rows}
\usage{
sf_lines_to_df(
  lines_sf,
  coord_colnames = c("shape_pt_lon", "shape_pt_lat"),
  remove_geometry = TRUE
)
}
\arguments{
\item{lines_sf}{sf object}

\item{coord_colnames}{names of the new columns (existing columns are overwritten)}

\item{remove_geometry}{remove sf geometry column?}
}
\description{
Adds the coordinates of an sf LINESTRING object as columns and rows
}
\keyword{internal}
