% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{sf_points_to_df}
\alias{sf_points_to_df}
\title{Adds the coordinates of an sf POINT object as columns}
\usage{
sf_points_to_df(
  pts_sf,
  coord_colnames = c("stop_lon", "stop_lat"),
  remove_geometry = TRUE
)
}
\arguments{
\item{pts_sf}{sf object}

\item{coord_colnames}{names of the new columns (existing columns are overwritten)}

\item{remove_geometry}{remove sf geometry column?}
}
\description{
Adds the coordinates of an sf POINT object as columns
}
\keyword{internal}
