% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{sf_to_json}
\alias{sf_to_json}
\title{Convert an sf object to a json list}
\usage{
sf_to_json(sf_obj, layer_name)
}
\arguments{
\item{sf_obj}{sf table}
}
\value{
json list
}
\description{
The sf object is written to a temporary file and re-read with jsonlite::read_json().
}
\keyword{internal}
