% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{shape_as_sf_linestring}
\alias{shape_as_sf_linestring}
\title{return an sf linestring with lat and long from gtfs}
\usage{
shape_as_sf_linestring(df)
}
\arguments{
\item{df}{dataframe from the gtfs shapes split() on shape_id}
}
\value{
st_linestring (sfr) object
}
\description{
return an sf linestring with lat and long from gtfs
}
\keyword{internal}
