% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins.R
\name{shape_for_route}
\alias{shape_for_route}
\title{Get a set of shapes for a route}
\usage{
shape_for_route(g1, select_route_id, select_service_id)
}
\arguments{
\item{a}{dataframe output by join_mega_and_hf_routes()}

\item{route_id}{the id of the route}

\item{service_id}{the service for which to get stops}
}
\value{
shapes for a route
}
\description{
Get a set of shapes for a route
}
\keyword{internal}
