% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-gtfs.R
\name{shape_to_sf_linestring}
\alias{shape_to_sf_linestring}
\title{return an sf linestring with lat and long from gtfs}
\usage{
shape_to_sf_linestring(df)
}
\arguments{
\item{df}{dataframe from the gtfsr shapes_df split() on shape_id}
}
\value{
st_linestring (sfr) object
}
\description{
return an sf linestring with lat and long from gtfs
}
