% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{shapes_as_sfg}
\alias{shapes_as_sfg}
\title{return an sf multilinestring with lat and long from gtfs for a route
TODO not used}
\usage{
shapes_as_sfg(df)
}
\arguments{
\item{df}{the shapes dataframe from a gtfs object}
}
\value{
a multilinestring simple feature geometry (sfg) for the routes
}
\description{
return an sf multilinestring with lat and long from gtfs for a route
TODO not used
}
\keyword{internal}
