% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{shapes_df_as_sfg}
\alias{shapes_df_as_sfg}
\title{return an sf multilinestring with lat and long from gtfs for a route}
\usage{
shapes_df_as_sfg(df)
}
\arguments{
\item{df}{the shapes_df dataframe from a gtfsr object}
}
\value{
a multilinestring simple feature geometry (sfg) for the routes
}
\description{
return an sf multilinestring with lat and long from gtfs for a route
}
\examples{
data(gtfs_obj)
shapes_sfg <- shapes_df_as_sfg(gtfs_obj$shapes_df)
plot(shapes_sfg[[1]])
}
