% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{stop_group_distances}
\alias{stop_group_distances}
\title{Calculates distances among stop within the same group column}
\usage{
stop_group_distances(gtfs_stops, by = "stop_name")
}
\arguments{
\item{gtfs_stops}{gtfs stops table}

\item{by}{group column, default: stop_name}
}
\value{
data.frame with one row per group containing a distance matrix (dists),
number of stop ids within that group (n_stop_ids) and distance summary values (dist_mean,, dist_median and dist_max).
}
\description{
By default calculates distances among stop_ids with the same name.
}
