% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{stops_df_as_sf}
\alias{stops_df_as_sf}
\title{This function is deprecated. Please use get_stop_geometry}
\usage{
stops_df_as_sf(stops_df)
}
\description{
This function is deprecated. Please use get_stop_geometry
}
