% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.R
\name{stops_for_service}
\alias{stops_for_service}
\title{Get a set of stops for a given set of service ids}
\usage{
stops_for_service(g1, select_service_id)
}
\arguments{
\item{g1}{gtfsr object}

\item{service_ids}{the service for which to get stops}
}
\value{
stops for a given service
}
\description{
Get a set of stops for a given set of service ids
}
\keyword{internal}
