% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.gtfs}
\alias{summary.gtfs}
\title{GTFS feed summary}
\usage{
\method{summary}{gtfs}(object, ...)
}
\arguments{
\item{object}{a gtfs_obj as read by read_gtfs()}

\item{...}{further specifications}
}
\description{
GTFS feed summary
}
