% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.tidygtfs}
\alias{summary.tidygtfs}
\title{GTFS feed summary}
\usage{
\method{summary}{tidygtfs}(object, ...)
}
\arguments{
\item{object}{a tidygtfs object as read by \code{\link[=read_gtfs]{read_gtfs()}}}

\item{...}{ignored for tidygtfs}
}
\value{
the tidygtfs object, invisibly
}
\description{
GTFS feed summary
}
