% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gtfs.R
\name{tfeeds_check}
\alias{tfeeds_check}
\title{Check HTTP status; stop if failure}
\usage{
tfeeds_check(req)
}
\arguments{
\item{req}{The result of an httr::GET}
}
\description{
Check HTTP status; stop if failure
}

