% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-gtfs-feeds.R
\name{tfeeds_get}
\alias{tfeeds_get}
\title{Set up GET for transitfeeds API}
\usage{
tfeeds_get(path, query, ..., version = "v1/", key = if (has_key())
  get_api_key())
}
\arguments{
\item{path}{Character containing the name of the API call (ex. getFeeds)}

\item{version}{Version of API as character that can be appended as part of path, defaults to v1}

\item{key}{API key, defaults to get_api_key()}
}
\value{
Result of httr::GET
}
\description{
Set up GET for transitfeeds API
}
\details{
See http://transitfeeds.com/api/ for available API calls
}

