% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-gtfs-feeds.R
\name{tfeeds_parse_getfeedlist}
\alias{tfeeds_parse_getfeedlist}
\title{Parse a gtfs feed list}
\usage{
tfeeds_parse_getfeedlist(req)
}
\arguments{
\item{req}{The result of a GET}
}
\value{
Dataframe of feeds
}
\description{
Parse a gtfs feed list
}
