% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gtfs.R
\name{tfeeds_parse_getlocation}
\alias{tfeeds_parse_getlocation}
\title{Extract location dataframe from transitfeeds getLocation API}
\usage{
tfeeds_parse_getlocation(req)
}
\arguments{
\item{req}{Response to getLocation httr::GET}
}
\value{
Dataframe of locations with id, descriptions, and lat/lng
}
\description{
Extract location dataframe from transitfeeds getLocation API
}

