% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-gtfs-feeds.R
\name{tfeeds_text}
\alias{tfeeds_text}
\title{Parse content as text}
\usage{
tfeeds_text(req)
}
\arguments{
\item{req}{The result of a httr::GET}
}
\value{
content parsed as text
}
\description{
Parse content as text
}

