% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfsio.R
\name{tidygtfs_to_gtfs}
\alias{tidygtfs_to_gtfs}
\title{Convert a tidygtfs object to a gtfs object (for gtfsio)}
\usage{
tidygtfs_to_gtfs(gtfs_obj)
}
\arguments{
\item{gtfs_obj}{gtfs feed (tidygtfs object)}
}
\value{
gtfs list
}
\description{
Convert a tidygtfs object to a gtfs object (for gtfsio)
}
\keyword{internal}
