% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytransit.R
\docType{package}
\name{tidytransit-package}
\alias{tidytransit}
\alias{tidytransit-package}
\title{tidytransit: Easily Install and Load packages to Work With the General Transit Feed Specification}
\description{
'tidytransit' is a set of packages that work in harmony
because they work with a common data specification, namely
the General Transit Feed Specification (GTFS). This
package is designed to make it easy to install and load multiple
'tidytransit' packages in a single step. Learn more about 'tidytransit'
at <https://r-transit.org>.
}
\author{
\strong{Maintainer}: Tom Buckley \email{tom@tbuckl.com}

}
\keyword{internal}
