% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{tidytransit_deps}
\alias{tidytransit_deps}
\title{List all tidytransit dependencies}
\usage{
tidytransit_deps(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
tidytransit packages.}
}
\description{
List all tidytransit dependencies
}
