% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{tidytransit_update}
\alias{tidytransit_update}
\title{Update tidytransit packages}
\usage{
tidytransit_update(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also check all dependencies of
tidytransit packages.}
}
\description{
This will check to see if all tidytransit packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
tidytransit_update()
}
}
