% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-gtfs-feeds.R
\name{tools_api_key}
\alias{tools_api_key}
\title{Tools to get, set, or check for API key}
\usage{
tools_api_key(value = NULL)
}
\arguments{
\item{value}{Character. A TransitFeed API Key.}
}
\description{
Tools to get, set, or check for API key
}

