% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trigger_suppressWarnings}
\alias{trigger_suppressWarnings}
\title{Used to trigger suppressWarnings or not}
\usage{
trigger_suppressWarnings(expr, quiet)
}
\arguments{
\item{expr}{expression to suppress}

\item{quiet}{Logical. Whether to suppress or not. \code{FALSE} by default.}
}
\description{
Used to trigger suppressWarnings or not
}

