% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gtfs.R
\name{unzip_gtfs}
\alias{unzip_gtfs}
\title{Unzip GTFS file and delete zip}
\usage{
unzip_gtfs(path, delete_zip = TRUE)
}
\arguments{
\item{path}{File path of zipped file}

\item{delete_zip}{Logical, whether to delete the zipped file after extraction.  Deletes by default.}
}
\description{
Unzip GTFS file and delete zip
}

