% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{unzip_gtfs_files}
\alias{unzip_gtfs_files}
\title{Unzip GTFS file and delete zip}
\usage{
unzip_gtfs_files(zipfile, delete_zip = FALSE, move_path = NULL,
  quiet = FALSE)
}
\arguments{
\item{zipfile}{path to zipped file}

\item{delete_zip}{Boolean. whether to delete the zipped file after extraction.  Deletes by default.}

\item{move_path}{Character. full file path to desire new location}

\item{quiet}{Boolean. Whether to output files found in folder.}
}
\value{
file path to directory with gtfs .txt files
}
\description{
Unzip GTFS file and delete zip
}
