% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_url}
\alias{valid_url}
\title{Used to check if a url is valid}
\usage{
valid_url(url, timeout = 5, test_url = TRUE, quiet = TRUE)
}
\arguments{
\item{url}{Character. URL.}

\item{timeout}{Integer. Seconds before timeout.}

\item{test_url}{Boolean. Whether to test if the url connects or not. FALSE by default (can take a while).}

\item{quiet}{Boolean. Whether to display output.}
}
\description{
Used to check if a url is valid
}
