% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_gtfs.R
\name{validate_feed}
\alias{validate_feed}
\title{Create validation list for a feed}
\usage{
validate_feed(feed, return_feed = TRUE)
}
\arguments{
\item{feed}{A GTFS feed list object with components agency_df, etc.}

\item{return_feed}{If TRUE, returns feed list with validate appended (true by default),
if FALSE, returns validate list only}
}
\value{
A feed list object with $validate added to the end of the list, or just validate list
}
\description{
Create validation list for a feed
}

