% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_gtfs}
\alias{validate_gtfs}
\title{Validation table that defines for every data frame within gtfs_obj
whether the file is required ('req'), optional ('opt') or an extra
file ('ext'). The same is applied for each field. The validation_status
column points problems with files/fiels.}
\usage{
validate_gtfs(gtfs_obj)
}
\description{
Validation table that defines for every data frame within gtfs_obj
whether the file is required ('req'), optional ('opt') or an extra
file ('ext'). The same is applied for each field. The validation_status
column points problems with files/fiels.
}
