% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_gtfs_structure}
\alias{validate_gtfs_structure}
\title{Create validation table for a gtfs_obj. It provides an overview of the structure of all files that were imported.}
\usage{
validate_gtfs_structure(gtfs_obj)
}
\arguments{
\item{gtfs_obj}{A GTFS list object with components agency_df, etc.}
}
\description{
Create validation table for a gtfs_obj. It provides an overview of the structure of all files that were imported.
}
