% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-gtfs-structure.R
\name{validate_vars_provided}
\alias{validate_vars_provided}
\title{Validate variables provided vs spec}
\usage{
validate_vars_provided(val_files, gtfs_obj)
}
\arguments{
\item{val_files}{The dataframe output of validate_files_provided for a single feed}

\item{gtfs_obj}{A 'gtfs' class object with components agency_df, etc.}
}
\value{
Dataframe with one record per file x column (columns in spec + any extra provided),
        with file and field specs and file and field provided statuses
}
\description{
Validate variables provided vs spec
}
