% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{agency_info}
\alias{agency_info}
\title{Create a text listing the first \code{max_agencies} agencies of the feed}
\usage{
agency_info(gtfs_obj, max_agencies = 3)
}
\arguments{
\item{gtfs_obj}{gtfs feed (tidygtfs object)}

\item{max_agencies}{max number of agencies to list before using "..."}
}
\value{
called for side effects
}
\description{
Create a text listing the first \code{max_agencies} agencies of the feed
}
\keyword{internal}
