% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-gtfs.R
\name{convert_gtfs_routes_to_sf}
\alias{convert_gtfs_routes_to_sf}
\title{Convert GTFS data to simple features}
\usage{
convert_gtfs_routes_to_sf(gtfs_obj)
}
\arguments{
\item{gtfs_obj}{gtfsr object}
}
\value{
a simple features \code{sf} dataframe with \code{MULTILINESTRING} geometry of GTFS routes
}
\description{
Convert routes to a simple features \code{sf} dataframe from a GTFS object.
}
\examples{
data(gtfs_obj)
library(sf)
gtfs_sf <- convert_gtfs_routes_to_sf(gtfs_obj)
plot(gtfs_sf[1])
}
