% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{feed_contains}
\alias{feed_contains}
\title{Returns TRUE if the given gtfs_obj contains the table. Used to check for
tidytransit's calculated tables in sublist (\code{gtfs_obj$.})}
\usage{
feed_contains(gtfs_obj, table_name)
}
\arguments{
\item{gtfs_obj}{gtfs feed (tidygtfs object)}

\item{table_name}{name of the table to look for, as string}
}
\description{
Returns TRUE if the given gtfs_obj contains the table. Used to check for
tidytransit's calculated tables in sublist (\code{gtfs_obj$.})
}
