% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gtfs.R
\name{filter_feedlist}
\alias{filter_feedlist}
\title{Filter a feedlist for just downloadable results}
\usage{
filter_feedlist(feedlist_df)
}
\arguments{
\item{feedlist_df}{A dataframe of feed metadata such as output from get_feedlist}
}
\value{
A dataframe of feed metadata for all feeds in input that are downloadable
}
\description{
Filter a feedlist for just downloadable results
}

