% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{get_feed}
\alias{get_feed}
\title{Download a zipped GTFS feed file from a url}
\usage{
get_feed(url, path = NULL, quiet = FALSE)
}
\arguments{
\item{url}{Character URL of GTFS feed.}

\item{path}{Character. Folder into which to put zipped file. If NULL, then save a tempfile}

\item{quiet}{Boolean. Whether to see file download progress. FALSE by default.}
}
\value{
File path
}
\description{
Download a zipped GTFS feed file from a url
}
