% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitfeeds.R
\name{get_feedlist}
\alias{get_feedlist}
\title{Get list of all available feeds from transitfeeds API}
\usage{
get_feedlist()
}
\value{
a data frame with the gtfs feeds on transitfeeds.
}
\description{
Get list of all available feeds from transitfeeds API
}
\examples{
\dontrun{
feedlist_df <- get_feedlist()
}
}
\seealso{
feedlist_df
}
