% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gtfs.R
\name{get_feedlist}
\alias{get_feedlist}
\title{Get list of available feeds from transitfeeds API}
\usage{
get_feedlist(location_id)
}
\value{
Result of GET
}
\description{
Get list of available feeds from transitfeeds API
}

