% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{gt_as_dt}
\alias{gt_as_dt}
\title{Get a dataframe with lubridate dates for the gtfs stop_times_df}
\usage{
gt_as_dt(stop_times_df)
}
\arguments{
\item{stop_times_df}{a gtfsr$stop_times_df dataframe}
}
\value{
an dataframe with arrival and departure time set to lubridate types
}
\description{
Get a dataframe with lubridate dates for the gtfs stop_times_df
}
\examples{
stop_times_df <- gt_as_dt(gtfs_obj$stop_times_df)
hours_v <- hour(stop_times_df$departure_time)
histogram(hours_v)
}
