% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{gtfs_as_sf}
\alias{gtfs_as_sf}
\title{Get common simple features (sf) for a gtfsr object}
\usage{
gtfs_as_sf(gtfs_obj)
}
\arguments{
\item{gtfs_obj}{a standard gtfsr object}
}
\value{
gtfs_obj a gtfsr object with a bunch of simple features tables
}
\description{
Get common simple features (sf) for a gtfsr object
}
