% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{gtfs_as_sf}
\alias{gtfs_as_sf}
\title{Add Simple Features for Stops and Routes to GTFS Object}
\usage{
gtfs_as_sf(gtfs_obj, quiet)
}
\arguments{
\item{gtfs_obj}{a standard tidytransit gtfs object}

\item{quiet}{boolean whether to print status messages}
}
\value{
gtfs_obj a tidytransit gtfs object with a bunch of simple features tables
}
\description{
Add Simple Features for Stops and Routes to GTFS Object
}
