% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtfs_obj}
\alias{gtfs_obj}
\title{Example GTFS data}
\description{
Data obtained from
\url{http://data.trilliumtransit.com/gtfs/duke-nc-us/duke-nc-us.zip}.
}
\seealso{
convert_gtfs_routes_to_sf
}
