% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{hhmmss_to_hms}
\alias{hhmmss_to_hms}
\title{Convert "HH:MM:SS" time strings to hms values
empty strings are converted to NA}
\usage{
hhmmss_to_hms(time_strings)
}
\arguments{
\item{time_strings}{char vector ("HH:MM:SS")}
}
\description{
Convert "HH:MM:SS" time strings to hms values
empty strings are converted to NA
}
\keyword{internal}
