% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{hhmmss_to_seconds}
\alias{hhmmss_to_seconds}
\title{Function to convert "HH:MM:SS" time strings to seconds.}
\usage{
hhmmss_to_seconds(hhmmss_str)
}
\arguments{
\item{hhmmss_str}{string}
}
\description{
Function to convert "HH:MM:SS" time strings to seconds.
}
\keyword{internal}
