% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-gtfs.R
\name{import_gtfs}
\alias{import_gtfs}
\title{Get a Dataframes of GTFS data.}
\usage{
import_gtfs(urls, quiet = FALSE)
}
\arguments{
\item{urls}{Character. url link to zip file}

\item{quiet}{Boolean. Whether to see file download progress and files extract. FALSE by default.}
}
\value{
Dataframes of GTFS data.
}
\description{
Get a Dataframes of GTFS data.
}

