% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-gtfs.R
\name{import_gtfs}
\alias{import_gtfs}
\title{Get a Dataframes of GTFS data.}
\usage{
import_gtfs(urls, timeout = 10)
}
\arguments{
\item{path}{Character. url link to zip file}
}
\value{
Dataframes of GTFS data.
}
\description{
Get a Dataframes of GTFS data.
}

