% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_gtfs}
\alias{import_gtfs}
\title{Get dataframes of GTFS data.}
\usage{
import_gtfs(path, data_dir = tempdir(), local = FALSE, quiet = FALSE)
}
\arguments{
\item{local}{Boolean. If the paths are searching locally or not. Default is FALSE (that is, urls).}

\item{quiet}{Boolean. Whether to see file download progress and files extract. FALSE by default.}

\item{paths}{Character. url links to zip files OR paths to local zip files. if to local path, then option \code{local} must be set to TRUE.}
}
\value{
Dataframes of GTFS data.
}
\description{
Get dataframes of GTFS data.
}
\examples{
\dontrun{
accra_gtfs <- import_gtfs("https://github.com/AFDLab4Dev/AccraMobility/raw/master/GTFS/GTFS_Accra.zip")
}
}
