% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{json_to_sf}
\alias{json_to_sf}
\title{Convert a json (read with jsonlite) to sf object}
\usage{
json_to_sf(json_list)
}
\arguments{
\item{json_list}{list as read by jsonlite::read_json (in gtfsio)}
}
\value{
sf object
}
\description{
The json object is written to a temporary file and re-read with sf::read().
}
\keyword{internal}
