% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-gtfs-networks.R
\name{map_gtfs_agency_routes}
\alias{map_gtfs_agency_routes}
\title{map all routes for an agency}
\usage{
map_gtfs_agency_routes(gtfs_obj, agency_id, include_stops = FALSE)
}
\arguments{
\item{gtfs_obj}{A GTFS list object with components agency_df, etc.}

\item{agency_id}{Character. Provide the ID of the agency whose routes are being mapped.}

\item{include_stops}{Boolean. Whether to layer on stops to the route shape. Default is FALSE}
}
\value{
Leaflet map object with all routes plotted for given agency ID.
}
\description{
map all routes for an agency
}

