% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/map-gtfs.R
\name{map_gtfs_route_shape}
\alias{map_gtfs_route_shape}
\title{map the shape, with stops, for a route}
\usage{
map_gtfs_route_shape(gtfs_obj, route_id, include_stops = TRUE)
}
\arguments{
\item{gtfs_obj}{A GTFS list object with components agency_df, etc.}

\item{route_id}{Character. A single ID for a route of interest.}

\item{include_stops}{Boolean. Whether to layer on stops to the route shape. Default is TRUE.}
}
\value{
Leaflet map object with all stop lat/long values plotted for a route.
}
\description{
map the shape, with stops, for a route
}

