% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-gtfs.R
\name{map_gtfs_stop}
\alias{map_gtfs_stop}
\title{map a single stop}
\usage{
map_gtfs_stop(gtfs_obj, stop_id)
}
\arguments{
\item{gtfs_obj}{A GTFS list object with components agency_df, etc.}

\item{stop_id}{Character. A single ID for a stop of interest.}
}
\value{
Leaflet map object with point plotted at stop lat/long value
}
\description{
map a single stop
}

