% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{na_to_empty_strings}
\alias{na_to_empty_strings}
\title{Convert NA values to empty strings ("")}
\usage{
na_to_empty_strings(gtfs_obj)
}
\arguments{
\item{gtfs_obj}{gtfs feed (tidygtfs object)}
}
\value{
a gtfs_obj where all NA strings in tables have been replaced with ""
}
\description{
\code{\link[=write_gtfs]{write_gtfs()}} converts \code{NA} to empty strings
}
\seealso{
\code{\link[=empty_strings_to_na]{empty_strings_to_na()}}
}
