% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tidygtfs}
\alias{plot.tidygtfs}
\title{Plot GTFS object routes and their frequencies}
\usage{
\method{plot}{tidygtfs}(x, ...)
}
\arguments{
\item{x}{a gtfs_obj as read by read_gtfs()}

\item{...}{further specifications}
}
\description{
Plot GTFS object routes and their frequencies
}
\examples{
\donttest{
local_gtfs_path <- system.file("extdata",
                              "google_transit_nyc_subway.zip",
                              package = "tidytransit")
nyc <- read_gtfs(local_gtfs_path)
plot(nyc)
}
}
