% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/problems.R
\docType{data}
\name{problems}
\alias{problems}
\title{Assign function environment to 'problems'}
\format{\preformatted{List of 30
 $ unrecognized_column                                 :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 10 24 12 2 31 9 10 12
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fe237864d90> 
 $ deprecated_column                                   :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 14 22 16 2 29 9 14 16
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fe237864d90> 
 $ duplicate_column                                    :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 18 21 20 2 28 9 18 20
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fe237864d90> 
 $ missing_value                                       :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 22 18 24 2 25 9 22 24
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fe237864d90> 
 $ invalid_value                                       :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 26 18 28 2 25 9 26 28
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fe237864d90> 
 $ invalid_float_value                                 :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 30 24 32 2 31 9 30 32
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fe237864d90> 
 $ invalid_non_negative_integer_value                  :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 34 39 36 2 46 9 34 36
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fe237864d90> 
 $ duplicate_id                                        :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 38 17 40 2 24 9 38 40
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fe237864d90> 
 $ invalid_agency_id                                   :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 42 22 44 2 29 9 42 44
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fe237864d90> 
 $ unused_stop                                         :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 46 16 48 2 23 9 46 48
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fe237864d90> 
 $ used_station                                        :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 50 17 52 2 24 9 50 52
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fe237864d90> 
 $ stop_too_far_from_parent_station                    :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 54 37 56 2 44 9 54 56
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fe237864d90> 
 $ stops_too_close                                     :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 58 20 60 2 27 9 58 60
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fe237864d90> 
 $ stations_too_close                                  :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 62 23 64 2 30 9 62 64
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fe237864d90> 
 $ different_station_too_close                         :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 66 32 68 2 39 9 66 68
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fe237864d90> 
 $ stop_too_far_from_shape_with_dist_traveled          :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 70 47 72 2 54 9 70 72
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fe237864d90> 
 $ expiration_date                                     :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 74 20 76 2 27 9 74 76
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fe237864d90> 
 $ future_service                                      :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 78 19 80 2 26 9 78 80
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fe237864d90> 
 $ date_outside_valid_range                            :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 82 29 84 2 36 9 82 84
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fe237864d90> 
 $ no_service_exceptions                               :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 86 26 88 2 33 9 86 88
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fe237864d90> 
 $ too_fast_travel                                     :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 90 20 92 2 27 9 90 92
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fe237864d90> 
 $ stop_with_multiple_route_types                      :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 94 35 96 2 42 9 94 96
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fe237864d90> 
 $ duplicate_trip                                      :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 98 19 100 2 26 9 98 100
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fe237864d90> 
 $ overlapping_trips_in_same_block                     :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 102 36 104 2 43 9 102 104
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fe237864d90> 
 $ transfer_distance_too_big                           :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 106 30 108 2 37 9 106 108
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fe237864d90> 
 $ transfer_walking_speed_too_fast                     :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 110 36 112 2 43 9 110 112
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fe237864d90> 
 $ other_problem                                       :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 114 18 116 2 25 9 114 116
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fe237864d90> 
 $ too_many_days_without_service                       :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 118 34 120 2 41 9 118 120
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fe237864d90> 
 $ minimum_transfer_time_set_with_invalid_transfer_type:function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 122 57 124 2 64 9 122 124
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fe237864d90> 
 $ too_many_consecutive_stop_times_with_same_time      :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 126 51 128 2 58 9 126 128
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fe237864d90> 
}}
\usage{
problems
}
\description{
Assign function environment to 'problems'
}
\keyword{datasets}

