% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problems.R
\docType{data}
\name{problems}
\alias{problems}
\title{Assign function environment to 'problems'}
\format{An object of class \code{list} of length 30.}
\usage{
problems
}
\description{
Assign function environment to 'problems'
}
\keyword{datasets}

