% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-gtfs-files.R
\name{read_sub_gtfs}
\alias{read_sub_gtfs}
\title{Function to read all files into dataframes}
\usage{
read_sub_gtfs(file_path, assign_envir)
}
\arguments{
\item{file_path}{Character file path}

\item{assign_envir}{Environment Object. Option of where to assign dataframes.}
}
\description{
Function to read all files into dataframes
}

