% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gtfs.R
\name{read_sub_gtfs}
\alias{read_sub_gtfs}
\title{Function to read all files into dataframes}
\usage{
read_sub_gtfs(file_path, assign_envir, problems = FALSE)
}
\arguments{
\item{file_path}{Character file path}

\item{assign_envir}{Environment Object. Option of where to assign dataframes.}

\item{problems}{Logical (default is FALSE). Option whether to return a dataframe of problems}
}
\description{
Function to read all files into dataframes
}

