% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read-gtfs-files.R
\name{read_sub_gtfs}
\alias{read_sub_gtfs}
\title{Function to read all files into dataframes}
\usage{
read_sub_gtfs(file_path, assign_envir, quiet = FALSE)
}
\arguments{
\item{file_path}{Character file path}

\item{assign_envir}{Environment Object. Option of where to assign dataframes.}

\item{quiet}{Boolean. Whether to output messages and files found in folder.}
}
\description{
Function to read all files into dataframes
}

