% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{routes_df_as_sf}
\alias{routes_df_as_sf}
\title{Make Routes into Simple Features Lines}
\usage{
routes_df_as_sf(gtfs_obj, route_ids = NULL, service_ids = NULL)
}
\arguments{
\item{gtfs_obj}{tidytransit gtfs object}

\item{route_ids}{select routes to convert to simple features}

\item{service_ids}{select service_ids to convert to simple features}
}
\value{
an sf dataframe for gtfs routes with a multilinestring column
}
\description{
Make Routes into Simple Features Lines
}
\examples{
data(gtfs_obj)
routes_sf <- routes_df_as_sf(gtfs_obj)
plot(routes_sf[1,])
}
