% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-gtfs.R
\name{routes_df_as_sf}
\alias{routes_df_as_sf}
\title{return an sf dataframe with from a gtfs object}
\usage{
routes_df_as_sf(gtfs_obj)
}
\arguments{
\item{gtfs_obj}{gtfsr object}
}
\value{
an sf dataframe for gtfs routes with a multilinestring column
}
\description{
return an sf dataframe with from a gtfs object
}
\examples{
url <- "http://data.trilliumtransit.com/gtfs/duke-nc-us/duke-nc-us.zip"
gtfs_obj <- url \%>\% import_gtfs(quiet=TRUE)
shapes_sf <- routes_df_as_sf(gtfs_obj$shapes_df)
plot(shapes_sf)
}
