% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-gtfs.R
\name{routes_df_as_sf}
\alias{routes_df_as_sf}
\title{Get a \code{sf} dataframe for gtfs routes}
\usage{
routes_df_as_sf(gtfs_obj)
}
\arguments{
\item{gtfs_obj}{gtfsr object}
}
\value{
an sf dataframe for gtfs routes with a multilinestring column
}
\description{
Get a \code{sf} dataframe for gtfs routes
}
\examples{
routes_sf <- routes_df_as_sf(gtfs_obj$shapes_df)
plot(routes_sf[1])
}
