% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{set_dates_services}
\alias{set_dates_services}
\title{Returns all possible date/service_id combinations as a data frame}
\usage{
set_dates_services(gtfs_obj)
}
\arguments{
\item{gtfs_obj}{gtfs feed (tidygtfs object)}
}
\value{
a date_service data frame
}
\description{
Returns all possible date/service_id combinations as a data frame
}
\keyword{internal}
