% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.R
\name{set_servicepattern}
\alias{set_servicepattern}
\title{Calculate servicepattern for the gtfs_obj}
\usage{
set_servicepattern(gtfs_obj)
}
\arguments{
\item{gtfs_obj}{feed}
}
\value{
modified gtfs_obj with added servicepattern list and a table linking trips and pattern (trip_servicepatterns)
}
\description{
Calculate servicepattern for the gtfs_obj
}
\keyword{internal}
