% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trips.R
\name{set_trippattern}
\alias{set_trippattern}
\title{Add trip pattern data frame to the gtfs object}
\usage{
set_trippattern(gtfs_obj)
}
\arguments{
\item{gtfs_obj}{gtfs feed}
}
\value{
gtfs_obj
}
\description{
Add trip pattern data frame to the gtfs object
}
