% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins.R
\name{shape_route_service}
\alias{shape_route_service}
\title{Join the shapes, trips and routes tables together - also checks on some potential errors in the data and warns accordingly}
\usage{
shape_route_service(gtfs_obj, route_ids = NULL, service_ids = NULL)
}
\arguments{
\item{gtfs_obj}{a gtfs object}

\item{route_ids}{the routes for which to join the tables together - required, but not sure why this can't just be any/all routes in routes_df}

\item{service_ids}{\itemize{
\item an optional filter for a certain service-default NULL
}}
}
\value{
shapes_routes_service - a dataframe in which routes, services, and shape_ids are all joined
}
\description{
Join the shapes, trips and routes tables together - also checks on some potential errors in the data and warns accordingly
}
\keyword{internal}
