% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{shapes_as_sf}
\alias{shapes_as_sf}
\title{Convert shapes into Simple Features Linestrings}
\usage{
shapes_as_sf(gtfs_shapes)
}
\arguments{
\item{gtfs_shapes}{a gtfs$shapes dataframe}
}
\value{
an sf dataframe for gtfs shapes
}
\description{
Convert shapes into Simple Features Linestrings
}
