% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{stops_as_sf}
\alias{stops_as_sf}
\title{Convert stops into Simple Features Points}
\usage{
stops_as_sf(stops, crs = NULL)
}
\arguments{
\item{stops}{a gtfs$stops dataframe}

\item{crs}{optional coordinate reference system (used by sf::st_transform) to transform
lon/lat coordinates}
}
\value{
an sf dataframe for gtfs routes with a point column
}
\description{
Convert stops into Simple Features Points
}
\examples{
data(gtfs_duke)
some_stops <- gtfs_duke$stops[sample(nrow(gtfs_duke$stops), 40),]
some_stops_sf <- stops_as_sf(some_stops)
plot(some_stops_sf)
}
