% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{stops_df_as_sf}
\alias{stops_df_as_sf}
\title{This function is deprecated. Please use get_stop_geometry
Make Stops into Simple Features Points}
\usage{
stops_df_as_sf(stops)
}
\arguments{
\item{stops}{a gtfsr$stops dataframe}
}
\value{
an sf dataframe for gtfs routes with a point column
}
\description{
This function is deprecated. Please use get_stop_geometry
Make Stops into Simple Features Points
}
\examples{
data(gtfs_obj)
some_stops <- gtfs_obj$stops[sample(nrow(gtfs_obj$stops), 40),]
some_stops_sf <- get_stop_geometry(some_stops)
plot(some_stops_sf)
}
