% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-gtfs-feeds.R
\name{tfeeds_get}
\alias{tfeeds_get}
\title{Set up GET for transitfeeds API}
\usage{
tfeeds_get(path, query, ..., version = "v1/", key = if (has_api_key())
  gtfs_api_key$get())
}
\arguments{
\item{path}{Character. Contains the name of the API call (ex. getFeeds)}

\item{query}{Character. Other query strings.}

\item{...}{any httr::GET options to pass-through}

\item{version}{Version of API as character that can be appended as part of path, defaults to v1}

\item{key}{API key, defaults to gtfs_api_key}
}
\value{
Result of httr::GET
}
\description{
Set up GET for transitfeeds API
}
\details{
See http://transitfeeds.com/api/ for available API calls
}
