% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-gtfs.R
\name{validate_files_provided}
\alias{validate_files_provided}
\title{For a single GTFS list object, check provided files}
\usage{
validate_files_provided(gtfs_obj)
}
\arguments{
\item{gtfs_obj}{A GTFS list object with components agency_df, etc.}
}
\value{
Dataframe will one row for all required and optional files per spec, plus one row for any other files provided (file),
        with an indication of these categories (spec), and a yes/no/empty status (provided_status)
}
\description{
For a single GTFS list object, check provided files
}

