% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-gtfs.R
\name{validate_gtfs_structure}
\alias{validate_gtfs_structure}
\title{Create validation list for a gtfs_obj. It provides an overview of the structure of all files that were imported.}
\usage{
validate_gtfs_structure(gtfs_obj, return_gtfs_obj = TRUE)
}
\arguments{
\item{gtfs_obj}{A GTFS list object with components agency_df, etc.}

\item{return_gtfs_obj}{If TRUE, returns gtfs_obj list with validate appended (true by default),
if FALSE, returns validate list only}
}
\value{
A gtfs_obj list object with $validate added to the end of the list, or just validate list
}
\description{
Create validation list for a gtfs_obj. It provides an overview of the structure of all files that were imported.
}

