% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_gtfs}
\alias{write_gtfs}
\title{Write a tidygtfs object to a zip file}
\usage{
write_gtfs(gtfs_obj, zipfile, compression_level = 9, as_dir = FALSE)
}
\arguments{
\item{gtfs_obj}{a gtfs feed object}

\item{zipfile}{path to the zip file the feed should be written to}

\item{compression_level}{a number between 1 and 9.9, passed to zip::zip}

\item{as_dir}{if TRUE, the feed is not zipped and zipfile is used as a directory path.
Files within the directory will be overwritten.}
}
\description{
Write a tidygtfs object to a zip file
}
\note{
Auxilliary tidytransit tables (e.g. \code{dates_services}) are not exported.
}
