% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_gtfs}
\alias{write_gtfs}
\title{Writes a gtfs object to a zip file. Calculated tidytransit tables and columns are not exported.}
\usage{
write_gtfs(gtfs_obj, zipfile, compression_level = 9)
}
\arguments{
\item{gtfs_obj}{a gtfs feed object}

\item{zipfile}{path to the zip file the feed should be written to}

\item{compression_level}{a number between 1 and 9.9, passed to zip::zip}
}
\description{
Writes a gtfs object to a zip file. Calculated tidytransit tables and columns are not exported.
}
